TLE-sender/src/requirements.txt
paramiko
sshtunnel
kafka-python==2.0.2